requirements.txt
fastmcp
langchain-chroma
langchain-huggingface
langchain-text-splitters
langchain-community
sentence-transformers
chromadb